refactor (redux): Change redux and some other components structure
