+++ package.json
@@ -7,7 +7,8 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "i": "yarn && cd ios && pod install" 
   },
   "dependencies": {
     "@react-navigation/native": "^6.0.10",
