+++ src/Navigation/screens.js
@@ -1,11 +1,11 @@
 import HomeScreen from '../Screens/HomeScreen';
 import UserProfileScreen from '../Screens/UserProfileScreen';
-import SeachContact from '../Screens/SeachContactButton';
+import SearchContact from '../Screens/SearchContact';
 import AddContact from '../Screens/AddContact';
 
 export default {
   HomeScreen,
   UserProfileScreen,
-  SeachContact,
+  SearchContact,
   AddContact,
 };
